fast-api/pyproject.toml
[project]
name = "fast-api"
version = "0.1.0"
description = "Python API for IA stuff"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.12",
  "google-adk[database]>=1.0.0",
  "google-cloud-aiplatform",
  "google-cloud-speech>=2.32.0",
  "google-cloud-texttospeech>=2.27.0",
  "google-generativeai",
  "ipython>=8.36.0",
  "litellm>=1.70.2",
  "python-dotenv>=1.1.0",
  "sentry-sdk>=2.27.0",
  "surrealdb>=1.0.4",
  "unicorn>=2.1.3",
  "unidecode>=1.3.8",
  "uvicorn>=0.34.2",
]

[project.optional-dependencies]
dev = ["debugpy>=1.6.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "features*"]
